libfru: fix build on MacOS
* remove explicit include of <endian.h>: it should be included in <sys/types.h>
* redefine endianness conversion functions
* rename uuid_t structure since there is same named type defined on
  MacOS
* add type casting for print statbuf.st_size - it also has type off_t

+++ frugen.c
@@ -314,7 +314,7 @@ void load_from_binary_file(const char *fname,
 		fatal("Cannot allocate buffer");
 	}
 
-	debug(2, "Reading the template file of size %lu...", statbuf.st_size);
+	debug(2, "Reading the template file of size %jd...", (intmax_t)statbuf.st_size);
 	off_t total_read = 0;
 	off_t bytes_read;
 	do
@@ -324,10 +324,10 @@ void load_from_binary_file(const char *fname,
 			fatal("Error (%d) reading template file: %s", errno, strerror(errno));
 		}
 		else {
-			debug(3, "Read %jd bytes from file (%jd/%zd)",
+			debug(3, "Read %jd bytes from file (%jd/%jd)",
 			      (intmax_t)bytes_read,
 			      (intmax_t)total_read,
-			      statbuf.st_size);
+			      (intmax_t)statbuf.st_size);
 		}
 		total_read += bytes_read;
 	} while(total_read < statbuf.st_size && bytes_read);
@@ -335,7 +335,7 @@ void load_from_binary_file(const char *fname,
 	if(total_read < statbuf.st_size) {
 		warn("File was shorter than expected (%jd / %jd)",
 		     (intmax_t)total_read,
-		     statbuf.st_size);
+		     (intmax_t)statbuf.st_size);
 	}
 	close(fd);
